requirements.txt
gradio>=3.50.2
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.26.0
# File processing - uncommented for document handling
pandas>=2.0.0
pypdf>=3.15.1
python-docx>=0.8.11
# Optional dependency for sample file creation
# reportlab>=4.0.5
# Database connectors (uncomment as needed)
# sqlalchemy>=2.0.0
# pymongo>=4.5.0
# psycopg2-binary>=2.9.7
# AI/ML libraries - choose based on your preferred model
# transformers>=4.35.0
# torch>=2.0.0
# sentence-transformers>=2.2.2
# langchain>=0.1.0
# langchain-community>=0.0.10gradio>=3.50.2
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.26.0
# File processing - uncommented for document handling
pandas>=2.0.0
pypdf>=3.15.1
python-docx>=0.8.11
# Other components (uncomment as needed)
# Database connectors
# sqlalchemy>=2.0.0
# pymongo>=4.5.0
# psycopg2-binary>=2.9.7
# AI/ML libraries - choose based on your preferred model
# transformers>=4.35.0
# torch>=2.0.0
# sentence-transformers>=2.2.2
# langchain>=0.1.0
# langchain-community>=0.0.10